Remove NoImplicit from scala-3
