+++ core/src/main/scala-2/spire/NoImplicit.scala
