+++ core/src/main/scala/spire/std/array.scala
@@ -17,8 +17,6 @@ package spire
 package std
 
 import spire.algebra._
-import spire.NoImplicit
-import scala.annotation.nowarn
 
 object ArraySupport {
   import spire.syntax.order._
@@ -125,23 +123,6 @@ object ArraySupport {
   }
 }
 
-trait ArrayInstances0 {
-  type NI0[A] = NoImplicit[VectorSpace[Array[A], A]]
-
-  implicit def ArrayCModule[@sp(Int, Long, Float, Double) A: NI0: ClassTag: CRing]: CModule[Array[A], A] =
-    new ArrayCModule[A]
-}
-
-trait ArrayInstances1 extends ArrayInstances0 {
-  type NI1[A] = NoImplicit[NormedVectorSpace[Array[A], A]]
-
-  implicit def ArrayVectorSpace[@sp(Int, Long, Float, Double) A: NI1: ClassTag: Field]: VectorSpace[Array[A], A] =
-    new ArrayVectorSpace[A]
-
-  implicit def ArrayEq[@sp A: Eq]: Eq[Array[A]] =
-    new ArrayEq[A]
-}
-
 trait ArrayInstances2 extends ArrayInstances1 {
   implicit def ArrayInnerProductSpace[@sp(Float, Double) A: Field: ClassTag]: InnerProductSpace[Array[A], A] =
     new ArrayInnerProductSpace[A]
@@ -158,34 +139,6 @@ trait ArrayInstances3 extends ArrayInstances2 {
 trait ArrayInstances extends ArrayInstances3 {
   implicit def ArrayMonoid[@sp A: ClassTag]: Monoid[Array[A]] =
     new ArrayMonoid[A]
-}
-
-@SerialVersionUID(0L)
-@nowarn
-final private class ArrayCModule[@sp(Int, Long, Float, Double) A: ClassTag: CRing](implicit
-  nvs: NoImplicit[VectorSpace[Array[A], A]]
-) extends CModule[Array[A], A]
-    with Serializable {
-  def scalar: CRing[A] = CRing[A]
-  def zero: Array[A] = new Array[A](0)
-  def negate(x: Array[A]): Array[A] = ArraySupport.negate(x)
-  def plus(x: Array[A], y: Array[A]): Array[A] = ArraySupport.plus(x, y)
-  override def minus(x: Array[A], y: Array[A]): Array[A] = ArraySupport.minus(x, y)
-  def timesl(r: A, x: Array[A]): Array[A] = ArraySupport.timesl(r, x)
-}
-
-@SerialVersionUID(0L)
-@nowarn
-final private class ArrayVectorSpace[@sp(Int, Float, Long, Double) A: ClassTag: Field](implicit
-  nnvs: NoImplicit[NormedVectorSpace[Array[A], A]]
-) extends VectorSpace[Array[A], A]
-    with Serializable {
-  def scalar: Field[A] = Field[A]
-  def zero: Array[A] = new Array[A](0)
-  def negate(x: Array[A]): Array[A] = ArraySupport.negate(x)
-  def plus(x: Array[A], y: Array[A]): Array[A] = ArraySupport.plus(x, y)
-  override def minus(x: Array[A], y: Array[A]): Array[A] = ArraySupport.minus(x, y)
-  def timesl(r: A, x: Array[A]): Array[A] = ArraySupport.timesl(r, x)
 }
 
 @SerialVersionUID(0L)
