+++ core/src/main/scala/spire/std/seq.scala
@@ -22,7 +22,6 @@ import scala.collection.Seq
 import scala.collection.SeqOps
 
 import spire.algebra._
-import spire.NoImplicit
 import scala.annotation.nowarn
 
 @SerialVersionUID(0L)
@@ -273,27 +272,6 @@ class SeqVectorOrder[A: Order, SA <: SeqOps[A, Seq, SA]](implicit scalar: Additi
 
     loop(x.iterator, y.iterator)
   }
-}
-
-trait SeqInstances0 {
-  @nowarn
-  implicit def SeqCModule[A, CC[A] <: SeqOps[A, Seq, CC[A]]](implicit
-    ring0: CRing[A],
-    cbf0: Factory[A, CC[A]],
-    ev: NoImplicit[VectorSpace[CC[A], A]]
-  ): SeqCModule[A, CC[A]] = new SeqCModule[A, CC[A]]
-}
-
-trait SeqInstances1 extends SeqInstances0 {
-  @nowarn
-  implicit def SeqVectorSpace[A, CC[A] <: SeqOps[A, Seq, CC[A]]](implicit
-    field0: Field[A],
-    cbf0: Factory[A, CC[A]],
-    ev: NoImplicit[NormedVectorSpace[CC[A], A]]
-  ): SeqVectorSpace[A, CC[A]] = new SeqVectorSpace[A, CC[A]]
-
-  implicit def SeqEq[A, CC[A] <: SeqOps[A, Seq, CC[A]]](implicit A0: Eq[A]): SeqEq[A, CC[A]] =
-    new SeqEq[A, CC[A]]
 }
 
 trait SeqInstances2 extends SeqInstances1 {
